updated pop-up recipe
Перенесла модальне вікно в окремий файл, додала подію на закриття модалки по кнопці ESC
