+++ src/css/modals/recipe.css
@@ -9,7 +9,7 @@
 	display: flex;
 	opacity: 0.5;
 }
-.modal{
+.popup_modal{
 	max-width: 530px;
 	border-radius: 15px;
 	background-color: white;
