+++ src/js/card_template.js
@@ -1,4 +1,5 @@
 import {getRecipeById} from "./api";
+import {popUpFunction} from "./modals/recipe";
 
 function createCardTemplate(data, list) {
    const markup = data.map(({ _id, preview, title, description, rating, area, ingredients, time }) =>
@@ -40,102 +41,5 @@ document.addEventListener('click', function(event) {
       onButtonClick(event);
   }
 });
-
-function popUpFunction(_id) {
-  const modal_popup = document.querySelector('.pop-up-recipe');
-  getRecipeById(_id).then((recipe) => {
-    console.log(recipe);
-
-    let current_tags = "";
-    for (let i = 0; i < recipe.tags.length; i++) {
-      if (recipe.tags[i] !== ""){
-        current_tags += '<a href="" class="modal_tag">#'+recipe.tags[i]+"</a>";
-      }
-    }
-
-    let current_ingredients = "";
-    for (let i = 0; i < recipe.ingredients.length; i++) {
-      current_ingredients += `<div class="modal_ingredient"><span>${recipe.ingredients[i].name}</span><span>${recipe.ingredients[i].measure}</span></div>`;
-    }
-
-    let current_rating = Math.round(recipe.rating);
-    let current_stars = "";
-    for (let i = 0; i < current_rating; i++) {
-      current_stars += `<svg xmlns="http://www.w3.org/2000/svg" width="14" height="13" viewBox="0 0 14 13" fill="none">
-        <path d="M6.04894 0.927052C6.3483 0.0057416 7.6517 0.00574088 7.95106 0.927052L8.79611 3.52786C8.92999 3.93989 9.31394 4.21885 9.74717 4.21885H12.4818C13.4505 4.21885 13.8533 5.45846 13.0696 6.02786L10.8572 7.63525C10.5067 7.8899 10.3601 8.34127 10.494 8.75329L11.339 11.3541C11.6384 12.2754 10.5839 13.0415 9.80017 12.4721L7.58779 10.8647C7.2373 10.6101 6.7627 10.6101 6.41222 10.8647L4.19983 12.4721C3.41612 13.0415 2.36164 12.2754 2.66099 11.3541L3.50604 8.75329C3.63992 8.34127 3.49326 7.8899 3.14277 7.63525L0.930391 6.02787C0.146677 5.45846 0.549452 4.21885 1.51818 4.21885H4.25283C4.68606 4.21885 5.07001 3.93989 5.20389 3.52786L6.04894 0.927052Z" fill="#EEA10C"/>
-      </svg>`
-    }
-    for (let i = 0; i < 5 - current_rating; i++) {
-      current_stars += `<svg xmlns="http://www.w3.org/2000/svg" width="14" height="13" viewBox="0 0 14 13" fill="none">
-      <path d="M6.04894 0.927052C6.3483 0.0057416 7.6517 0.00574088 7.95106 0.927052L8.79611 3.52786C8.92999 3.93989 9.31394 4.21885 9.74717 4.21885H12.4818C13.4505 4.21885 13.8533 5.45846 13.0696 6.02786L10.8572 7.63525C10.5067 7.8899 10.3601 8.34127 10.494 8.75329L11.339 11.3541C11.6384 12.2754 10.5839 13.0415 9.80017 12.4721L7.58779 10.8647C7.2373 10.6101 6.7627 10.6101 6.41222 10.8647L4.19983 12.4721C3.41612 13.0415 2.36164 12.2754 2.66099 11.3541L3.50604 8.75329C3.63992 8.34127 3.49326 7.8899 3.14277 7.63525L0.930391 6.02787C0.146677 5.45846 0.549452 4.21885 1.51818 4.21885H4.25283C4.68606 4.21885 5.07001 3.93989 5.20389 3.52786L6.04894 0.927052Z" fill="grey"/>
-    </svg>`
-    }
-
-    let current_youtube = "";
-    if(recipe.youtube){
-      current_youtube = recipe.youtube.replace('watch?v=','embed/');
-    }
-
-    const transformedRecipe = `
-    <div id="bg_modal" class="bg_modal"></div>
-    <div class="modal">
-      <a id="pop-up-close" class="modal_close">
-        <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none">
-          <path d="M18 6L6 18" stroke="black" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/>
-          <path d="M6 6L18 18" stroke="black" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/>
-        </svg>
-      </a>
-  
-      <div class="modal_container">
-        <h4 class="modal_title">${recipe.title}</h4>
-        <iframe class="modal_video" src="${current_youtube}" referrerpolicy="no-referrer" ></iframe>
-        <div class="modal_line">
-          <div class="modal_tags">
-            ${current_tags}
-          </div>
-          <div class="modal_right">
-            <div class="modal_rating">
-              <span>${recipe.rating}</span>
-              <div class="modal_rating_stars">
-                ${current_stars}
-              </div>
-            </div>
-            <div class="modal_time">${recipe.time} min</div>
-          </div>
-        </div>
-  
-        <div class="modal_ingredients">
-          ${current_ingredients}
-        </div>
-        <div class="modal_description">${recipe.description}</div>
-        <div class="modal_buttons">
-          <div class="modal_button modal_favourite">
-            <button data-id="${recipe._id}">Add to favourite</button>
-          </div>
-          <div class="modal_button modal_rating_button">
-            <button data-id="${recipe._id}">Give a rating</button>
-          </div>
-        </div>
-      </div>
-    </div>
-    `;
-    modal_popup.innerHTML = transformedRecipe;
-
-
-    document.getElementById("pop-up-close").addEventListener("click", function () {
-      const popUpRecipes = document.querySelectorAll(".pop-up-recipe");
-      popUpRecipes.forEach(function (recipe) {
-        recipe.innerHTML = "";
-      });
-    });
-
-    document.getElementById("bg_modal").addEventListener("click", function () {
-      const popUpRecipes = document.querySelectorAll(".pop-up-recipe");
-      popUpRecipes.forEach(function (recipe) {
-        recipe.innerHTML = "";
-      });
-    });
-  })
-}
 
 export {createCardTemplate}
